pyproject.toml
[tool.poetry]
name = "animalmassage"
version = "0.1.0"
description = ""
authors = ["Eddy <eddy15201@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
isort = "^5.11.4"
black = "^22.12.0"
fastapi = "^0.89.0"
httpx = "^0.23.1"
pydantic = {version = "^1.10.2", extras = ["email"]}
pytest = "7.2.2"
pytest-cov = "^4.0.0"
uvicorn = "^0.20.0"
google-auth = "^2.16.3"
requests = "^2.28.2"
testcontainers = "3.4.2"
sqlalchemy = "1.4.45"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
